refactor(reservation): '예약 현황' 상태 관리, 액션 핸들러 hook 분리
rel: #88
